requirements.txt
pbr
Flask
python-freshdesk
keystoneauth1
oslo.config
oslo.context
oslo.log
oslo.messaging
oslo.middleware
python-keystoneclient
python-novaclient
python-neutronclient
python-glanceclient
python-cinderclient
gunicorn
